requirements.txt
tiktoken
emoji
openai
google-genai
scikit-learn
numpy
python-dotenv
